[MAJOR] Fix insert/delete logic
Insert/delete logic was flawed in the sense that the pointer insert
returned could not be safely used to delete/read the element later.
Insert function returned a pointer to an element in the heap array
which wasn't updated if the element it pointed at changed its position
or if a realloc happened and its address changed.

Fix this similarly to how doubly-linked list Linux implementation does
it: add a structure that must be encapsulated in the binary-heap elements
data (struct binheap_entry).

Also:
    # switch to function inlining instead of macros
	# add Eclipse Makefile project files
	    (you need Eclipse CDT and make+gcc+gdb in PATH to run it)

Signed-off-by: Corneliu Zuzu <zuzelzzz@gmail.com>